Loop through all resolved addresses in virNetSocketNewListenTCP
Currently virNetSocketNewListenTCP bails out early under the following
conditions:
- the hostname resolves to at least one IPv4 and at least one IPv6
  address
- the local interfaces have that one IPv4 address assigned, but not any
  of the IPv6 addresses
- the local interfaces have just IPv6 link-local addresses

In this case the resolver returns not only the IPv4 addreses but also
IPv6. Binding the IPv6 address will obviously fail. But this terminates
the entire loop, even if binding to IPv4 succeeded.

To fix this error, just keep going and loop through all returned
addresses. In case none of the attempts to bind to some address
succeeded, try to report the appropriate error.

Signed-off-by: Olaf Hering <olaf@aepfle.de>
Message-Id: <20180515082026.13928-1-olaf@aepfle.de>