log: actually do substring matches with fnmatch
Historically we matched log filters with strstr(), and when switching to
fnmatch in cbb0fd3cfdc287f6f4653ef1f04a7cfb2ea51b27, it was stated that
we would continue to match substrings, with "foo" being equivalent to
"*foo*". Unfortuntely I forget to provide the code to actually make that
happen. This fixes it to prepend and append "*". We don't bother to
check if the pattern already has a leading/trailing '*', because
"**foo**" will match the same as "*foo*".

Signed-off-by: Daniel P. Berrangé <berrange@redhat.com>
Message-Id: <20180516111716.5879-1-berrange@redhat.com>
Reviewed-by: Erik Skultety <eskultet@redhat.com>