storage_util: Alter qemu storage encryption arguments
https://bugzilla.redhat.com/show_bug.cgi?id=1526382

As of QEMU 2.9, qemu-img has enforced using the "encrypt.key-secret"
in order to create a qcow[2] encrypted volume. Thus, the existing code
to create an encrypted volume using qcow[2] encryption techniques will
fail, such as :

  $ qemu-img create -f qcow2 -b /dev/null \
        -o backing_fmt=raw,encryption=on \
        demo.tmp 5242880K
  Formatting 'demo.tmp', fmt=qcow2 size=5368709120 backing_file=/dev/null
  backing_fmt=raw encryption=on cluster_size=65536 lazy_refcounts=off
  refcount_bits=16
  qemu-img: demo.tmp: Parameter 'encrypt.key-secret' is required for cipher
  $

This patch will resolve this by adding the correct parameters for
the creation. The new format of parameters roughly follows that of
LUKS encryption model with a few minor differences:

   1. Usage of "encrypt.key-secret=$alias" instead of just plain
      "key-secret=$alias" as the parameter.
   2. Usage of "encrypt.format=aes" instead of "encryption=on"

The result is the following command syntax for the same example:

  $ qemu-img create -f qcow2 -b /dev/null \
        --object secret,id=OtherDemo.img_encrypt0,file=/path/to/secretFile \
        -o encrypt.format=aes,encrypt.key-secret=OtherDemo.img_encrypt0 \
        demo.tmp 5242880K
  Formatting 'test.img', fmt=qcow2 size=5368709120 backing_file=/dev/null
  backing_fmt=raw encrypt.format=aes encrypt.key-secret=sec0 cluster_size=65536
  lazy_refcounts=off refcount_bits=16
  $

Thus this patch removes the LUKS specific checks in a few places and
alters the algorithms as necessary in order to allow either form of
encryption.

The storagevolxml2argvtest.c test is adjusted to pass a dummy path to
the secret file and the outputs adjusted to illustrate the new format
for the various arguments.

This patch requires usage of the secrets object and model. There is no
plan for backwards compatibility for qcow[2] encryption. The desire is
to move towards usage of LUKS encryption anyway.

NB: Although the qemu-img convert examples change in the test output,
they are essentially still broken (they wouldn't work before this patch
either for the same reasons create fails). A follow-up patch will alter
the algorithm and syntax.

Signed-off-by: John Ferlan <jferlan@redhat.com>
Message-Id: <20180516122931.30854-2-jferlan@redhat.com>