storage_util: Don't assume "luks" encryption for resize
Similar to encrypted image creation/conversion resizing the
volume requires providing different parameters for luks and
qcow[2] encryption.

Alter storageBackendResizeQemuImgImageOpts to take the @type
parameter filled in during storageBackendResizeQemuImg to either
the current type or "luks" for a RAW image and use that to
generate the "driver=%s" value and to determine whether to use
"encrypt.key-secret" or "key-secret".

Signed-off-by: John Ferlan <jferlan@redhat.com>
Message-Id: <20180516122931.30854-4-jferlan@redhat.com>