qemu: Move checks for SMM from command-line creation into validation phase
We are still hoping all of such checks will be moved there and this is one small
step in that direction.

One of the things that this is improving is the error message you get when
starting a domain with SMM and i440fx, for example.  Instead of saying that the
QEMU binary doesn't support that option, we correctly say that it is only
supported with q35 machine type.

Signed-off-by: Martin Kletzander <mkletzan@redhat.com>
Message-Id: <85e343de944def6154c599ecda53bc70688ab60a.1526913931.git.mkletzan@redhat.com>