xen_common: convert to typesafe virConf acessors
There are still a few places using virConfGetValue(), checking for the
specific type of the pointers and so on.

Those places are not going to be converted as:
- Using virConfGetValue*() would trigger virReportError() in the current
code, which would cause, at least, some misleading messages for whoever
has to debug this code path;

- Expanding virConfValue*() to support strings as other types (for
instance, as boolean or long) does not seem to be the safest path to
take.

Signed-off-by: Fabiano Fidêncio <fabiano@fidencio.org>
Message-Id: <20180527222826.18525-7-fabiano@fidencio.org>