conf: domain: Remove code accessing 'bootHash' from the post-parse infrestructure
There is only one block accessing it. Removing it is necessary so that
post parse callbacks can later be invoked after the hypervisor private
data callback so that e.g. qemuCaps are properly filled for the
postparse callbacks.

As the function signature of virDomainDefPostParseInternal does not
differ from virDomainDefPostParse now, the wrapper can be dropped.

Signed-off-by: Peter Krempa <pkrempa@redhat.com>
Message-Id: <8b4627fedd199117af1ab46f5562e3838679357e.1527500017.git.pkrempa@redhat.com>