qemu: domain: Regenerate alias for the TLS x509 credential object
When restarting libvirt would previously lose the alias of the x509
certificate object. Upon unplug we would then not delete the
corresponding objects.

Restore the alias if we know it shoudl be there.

Luckily for disks we don't support encrypted TLS nevironment, so there's
no need to regenerate the 'seceret' alias for decrypting.

Signed-off-by: Peter Krempa <pkrempa@redhat.com>
Message-Id: <07ad650581506ba8f2afd4bfa81e0e07ccf0ede9.1527683836.git.pkrempa@redhat.com>
Reviewed-by: J�n Tomko <jtomko@redhat.com>