qemu: domain: aggregate setup of disk drive options for -drive
When using blockdev the approach to base aliases will change. Add a
helper function that will aggregate all code which needs to be called
with the disk alias for the -drive to setup internal data.

qemuDomainSecretDiskPrepare wrapper is no longer necessary as the
contents were moved to a function which is designed to use the old
aliases.

Signed-off-by: Peter Krempa <pkrempa@redhat.com>
Message-Id: <e4905aa84aa1d02ade9399b58aff017561b9eff3.1527683836.git.pkrempa@redhat.com>
Reviewed-by: J�n Tomko <jtomko@redhat.com>