conf: Introduce new <hostdev> attribute 'display'
QEMU introduced a new type of display for mediated devices using
vfio-pci backend which controls whether a mediated device can be used as
a native rendering device as an alternative to an emulated video device.
This patch adds the necessary bits to domain config handling in order to
expose this feature.
It's worth noting that even though QEMU supports and defaults to the
value 'auto', this patch only introduces values 'on' and 'off'
defaulting to 'off' (for safety), since there's no convenient way for
libvirt to come up with relevant defaults based on the fact, that libvirt
doesn't know whether the underlying device uses dma-buf or vfio region
mapping.

Signed-off-by: Erik Skultety <eskultet@redhat.com>
Message-Id: <edf0cb829dada07aae850e3bf41e2099791d93d4.1527685754.git.eskultet@redhat.com>