qemu: domain: Forbid storage with old QCOW2 encryption
The encryption was buggy and qemu actually dropped it upstream. Forbid
it for all versions since it would cause other problems too.

Problems with the old encryption include weak crypto, corruption of
images with blockjobs and a lot of usability problems.

This requires changing of the encryption type for the encrypted disk
tests.

Signed-off-by: Peter Krempa <pkrempa@redhat.com>
Message-Id: <e6a110fdfb9b99232826a3d920c7a92f06a119e5.1527854576.git.pkrempa@redhat.com>
Reviewed-by: J�n Tomko <jtomko@redhat.com>