qemu: conf: Remove 'allow_disk_format_probing' config option
The option is insecure and it has been long enough for users to migrate
their disk files to use explicit format. Drop the option and related
code.

The config parser still parses it and rejects statup if it's still
present in the config in enabled state.

The augeas lens is also kept so that users can disable it.

Signed-off-by: Peter Krempa <pkrempa@redhat.com>
Message-Id: <407ab3e587967a095137b7731596045cefc6a923.1528102541.git.pkrempa@redhat.com>
Reviewed-by: J�n Tomko <jtomko@redhat.com>