DO_NOT_APPLY: HACK: Regenerate qemucapabilitiestest data
Use the LD_PRELOAD .so library used in 'qemucapsprobe' to regenerate the
files. The library is instrumented to write the replies to 'caps.json'
and instrumentation in the test move it to the correct place.

Use 'fix-files.sh' to do so as it's using a static variable to
track the initial newline.
Message-Id: <110fa6b522fff3e8bf18a55717d7b684e72bbf61.1528120448.git.pkrempa@redhat.com>