qemu: Move checks for SMM from command-line creation into validation phase
We are still hoping all of such checks will be moved there and this is one small
step in that direction.

One of the things that this is improving is the fact that instead of error
message *that was wrong) you get when starting a domain with SMM and i440fx we
allow the setting to go through.  SMM option exists and makes sense on i440fx as
well (basically whenever that _SMM_OPT capability is set).

Signed-off-by: Martin Kletzander <mkletzan@redhat.com>
Message-Id: <ab19b26d50b8ea768cfeaa6d2dd88e69730163cf.1528360378.git.mkletzan@redhat.com>