Move host cache handling from conf/ to util/
Introduce virCache{Info,Bank} as structures describing host cache information
including the ways they can be managed by resctrl.  This has couple of
advantages.  First and foremost the virResctrlAllocCreate() is no longer
dependent on a data structure initialized in another part of the code and the
data is automatically fresh.  This couldn't be done previously since the code
initializing that was in conf/.  And it also makes more sense for functions that
parse sysfs files to be in util/.  Last, but not least, the code doesn't
duplicate any data or keep around stale information that needs to be
recalculated anyway.

Signed-off-by: Martin Kletzander <mkletzan@redhat.com>
Message-Id: <eb48791fdd3d62c4b172df2b4a985705b1b5ba0a.1528377367.git.mkletzan@redhat.com>