docker: Commit initial Dockerfiles
These have been generated from the build dependency data
available in the libvirt-jenkins-ci repository, and will
be refreshed periodically to keep them in sync the same
way we've updated .travis.yml so far; my guess, based on
that effort, is that we'll need to do so about once per
release.

Automated builds will be set up on Docker Hub so that
changes to the Dockerfiles will cause the images to be
regenerated, and with that in place (a subset of) the
resulting images will be used in the Travis CI pipeline,
as well of course as being available to developers for
testing and debugging purposes.

Signed-off-by: Andrea Bolognani <abologna@redhat.com>
Message-Id: <20180615070656.7085-2-abologna@redhat.com>