qemu_domain: Introduce qemuDomainObjBeginJobNowait
The aim of this API is to allow the caller do best effort. Some
functions can work even when acquiring the job fails (e.g.
qemuConnectGetAllDomainStats()). But what they can't bear is
delay if they have to wait up to 30 seconds for each domain that
is processing some other job.

Signed-off-by: Michal Privoznik <mprivozn@redhat.com>
Message-Id: <9afe290ab75b0a9830ea3b9bdd858d3609b1e9a1.1529050509.git.mprivozn@redhat.com>
Reviewed-by: John Ferlan <jferlan@redhat.com>